requirements.txt
asgiref==3.6.0
boto3==1.26.137
botocore==1.29.137
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-filter==23.2
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
jmespath==1.0.1
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
gunicorn==20.1.0
